rcreate a single form tag
