+++ .prettierrc.js
@@ -3,4 +3,5 @@ module.exports = {
     trailingComma: 'all',
     jsxSingleQuote: true,
     printWidth: 120,
+    tabWidth: 2,
   };
