+++ src/components/login.tsx
@@ -1,57 +1,32 @@
 import React, { CSSProperties } from 'react';
 
 export class EmailInput extends React.Component {
-    render() {
-        return (
-            <div>
-                <form style= {inputStyles}>
-                    E-mail
-                    <label>
-                        <input type="text"name= "email" />
-                    </label>
-                </form>
-            </div>
-        );
-    }
+  render() {
+    return (
+      <label>
+        <input type='text' name='email' />
+      </label>
+    );
+  }
 }
-
 
 export class PasswordInput extends React.Component {
-    render() {
-        return (
-            <div>
-                <form style={inputStyles}>
-                    Senha
-                    <label>
-                        <input type="text"name= "password" />
-                    </label>
-                </form>
-            </div>
-        );
-    }
+  render() {
+    return (
+      <label>
+        <input type='text' name='password' />
+      </label>
+    );
+  }
 }
-
 
 export class SubmitButton extends React.Component {
-    render () {
-        return (
-            <div>
-                <button style={buttonStyles}>
-                    Entrar
-                </button>
-            </div>
-        )
-    }
+  render() {
+    return <button style={buttonStyles}>Entrar</button>;
+  }
 }
 
-const inputStyles: CSSProperties = {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    margin: '20px'
-}
-
-const buttonStyles: CSSProperties ={
-    backgroundColor: 'blue',
-    color: 'white'
-}
+const buttonStyles: CSSProperties = {
+  backgroundColor: 'blue',
+  color: 'white',
+};
